refactor: move sync orchestrator to services
